Switches HAL to fixed length signed integers, and adds our own HAL_Bool Type (#155)
* Switches HAL to fixed length signed integers, and adds our own HAL_Bool type

* Replaces HAL Floats with Doubles

Doubles are just as fast as floats with optimizations turned on, so
switches to all doubles. All made doubles for consistency.

* Prepends HAL/ to HAL include files. Also fixes some range errors

+++ hal/include/HAL/AnalogGyro.h
@@ -9,7 +9,7 @@
 
 #include <stdint.h>
 
-#include "HAL/Handles.h"
+#include "HAL/Types.h"
 
 extern "C" {
 HAL_GyroHandle HAL_InitializeAnalogGyro(HAL_AnalogInputHandle handle,
@@ -17,17 +17,17 @@ HAL_GyroHandle HAL_InitializeAnalogGyro(HAL_AnalogInputHandle handle,
 void HAL_SetupAnalogGyro(HAL_GyroHandle handle, int32_t* status);
 void HAL_FreeAnalogGyro(HAL_GyroHandle handle);
 void HAL_SetAnalogGyroParameters(HAL_GyroHandle handle,
-                                 float voltsPerDegreePerSecond, float offset,
-                                 uint32_t center, int32_t* status);
+                                 double voltsPerDegreePerSecond, double offset,
+                                 int32_t center, int32_t* status);
 void HAL_SetAnalogGyroVoltsPerDegreePerSecond(HAL_GyroHandle handle,
-                                              float voltsPerDegreePerSecond,
+                                              double voltsPerDegreePerSecond,
                                               int32_t* status);
 void HAL_ResetAnalogGyro(HAL_GyroHandle handle, int32_t* status);
 void HAL_CalibrateAnalogGyro(HAL_GyroHandle handle, int32_t* status);
-void HAL_SetAnalogGyroDeadband(HAL_GyroHandle handle, float volts,
+void HAL_SetAnalogGyroDeadband(HAL_GyroHandle handle, double volts,
                                int32_t* status);
-float HAL_GetAnalogGyroAngle(HAL_GyroHandle handle, int32_t* status);
+double HAL_GetAnalogGyroAngle(HAL_GyroHandle handle, int32_t* status);
 double HAL_GetAnalogGyroRate(HAL_GyroHandle handle, int32_t* status);
-float HAL_GetAnalogGyroOffset(HAL_GyroHandle handle, int32_t* status);
-uint32_t HAL_GetAnalogGyroCenter(HAL_GyroHandle handle, int32_t* status);
+double HAL_GetAnalogGyroOffset(HAL_GyroHandle handle, int32_t* status);
+int32_t HAL_GetAnalogGyroCenter(HAL_GyroHandle handle, int32_t* status);
 }
