+++ hal/lib/athena/AnalogInternal.h
@@ -19,7 +19,7 @@ namespace hal {
 constexpr int32_t kTimebase = 40000000;  ///< 40 MHz clock
 constexpr int32_t kDefaultOversampleBits = 0;
 constexpr int32_t kDefaultAverageBits = 7;
-constexpr float kDefaultSampleRate = 50000.0;
+constexpr double kDefaultSampleRate = 50000.0;
 static const uint32_t kAccumulatorChannels[] = {0, 1};
 
 extern tAI* analogInputSystem;
